+++ common/src/main/java/de/maxhenkel/voicechat/voice/client/ClientManager.java
@@ -10,6 +10,7 @@ import de.maxhenkel.voicechat.macos.jna.avfoundation.AVAuthorizationStatus;
 import de.maxhenkel.voicechat.net.NetManager;
 import de.maxhenkel.voicechat.net.RequestSecretPacket;
 import de.maxhenkel.voicechat.net.SecretPacket;
+import de.maxhenkel.voicechat.voice.server.Server;
 import io.netty.channel.local.LocalAddress;
 import net.minecraft.ChatFormatting;
 import net.minecraft.Util;
@@ -44,6 +45,7 @@ public class ClientManager {
 
         ClientCompatibilityManager.INSTANCE.onJoinWorld(this::onJoinWorld);
         ClientCompatibilityManager.INSTANCE.onDisconnect(this::onDisconnect);
+        ClientCompatibilityManager.INSTANCE.onPublishServer(this::onPublishServer);
 
         ClientCompatibilityManager.INSTANCE.onVoiceChatConnected(connection -> {
             if (client != null) {
@@ -120,6 +122,29 @@ public class ClientManager {
             client = null;
         }
         ClientCompatibilityManager.INSTANCE.emitVoiceChatDisconnectedEvent();
+    }
+
+    private void onPublishServer(int port) {
+        Server server = Voicechat.SERVER.getServer();
+        if (server == null) {
+            return;
+        }
+        try {
+            Voicechat.LOGGER.info("Changing voice chat port to {}", port);
+            server.changePort(port);
+            ClientVoicechat client = ClientManager.getClient();
+            if (client != null) {
+                ClientVoicechatConnection connection = client.getConnection();
+                if (connection != null) {
+                    Voicechat.LOGGER.info("Force disconnecting due to port change");
+                    connection.disconnect();
+                }
+            }
+            NetManager.sendToServer(new RequestSecretPacket(Voicechat.COMPATIBILITY_VERSION));
+            Minecraft.getInstance().gui.getChat().addMessage(new TranslatableComponent("message.voicechat.server_port", server.getPort()));
+        } catch (Exception e) {
+            Voicechat.LOGGER.error("Failed to change voice chat port: {}", e.getMessage());
+        }
     }
 
     public void checkMicrophonePermissions() {