+++ fabric/src/main/java/de/maxhenkel/voicechat/mixin/IntegratedServerMixin.java
@@ -1,30 +1,38 @@
 package de.maxhenkel.voicechat.mixin;
 
-import de.maxhenkel.voicechat.Voicechat;
-import de.maxhenkel.voicechat.voice.server.Server;
-import net.minecraft.client.Minecraft;
+import com.mojang.authlib.GameProfileRepository;
+import com.mojang.authlib.minecraft.MinecraftSessionService;
+import com.mojang.datafixers.DataFixer;
+import de.maxhenkel.voicechat.events.PublishServerEvents;
 import net.minecraft.client.server.IntegratedServer;
-import net.minecraft.network.chat.TranslatableComponent;
+import net.minecraft.server.MinecraftServer;
+import net.minecraft.server.WorldStem;
+import net.minecraft.server.level.progress.ChunkProgressListenerFactory;
+import net.minecraft.server.packs.repository.PackRepository;
+import net.minecraft.server.players.GameProfileCache;
 import net.minecraft.world.level.GameType;
+import net.minecraft.world.level.storage.LevelStorageSource;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 import javax.annotation.Nullable;
+import java.net.Proxy;
 
 @Mixin(IntegratedServer.class)
-public class IntegratedServerMixin {
+public abstract class IntegratedServerMixin extends MinecraftServer {
+
+    public IntegratedServerMixin(Thread thread, LevelStorageSource.LevelStorageAccess levelStorageAccess, PackRepository packRepository, WorldStem worldStem, Proxy proxy, DataFixer dataFixer, @org.jetbrains.annotations.Nullable MinecraftSessionService minecraftSessionService, @org.jetbrains.annotations.Nullable GameProfileRepository gameProfileRepository, @org.jetbrains.annotations.Nullable GameProfileCache gameProfileCache, ChunkProgressListenerFactory chunkProgressListenerFactory) {
+        super(thread, levelStorageAccess, packRepository, worldStem, proxy, dataFixer, minecraftSessionService, gameProfileRepository, gameProfileCache, chunkProgressListenerFactory);
+    }
 
     @Inject(method = "publishServer", at = @At(value = "RETURN"))
     public void publishServer(@Nullable GameType gameType, boolean cheats, int port, CallbackInfoReturnable<Boolean> callbackInfo) {
         if (!callbackInfo.getReturnValue()) {
             return;
         }
-        Server server = Voicechat.SERVER.getServer();
-        if (server != null) {
-            Minecraft.getInstance().gui.getChat().addMessage(new TranslatableComponent("message.voicechat.server_port", server.getPort()));
-        }
+        PublishServerEvents.SERVER_PUBLISHED.invoker().accept(port);
     }
 
 }