+++ fabric/changelog.md
@@ -1,2 +1,2 @@
-- Handle uncaught exceptions when failing to load natives
-- Updated brazilian portuguese translation
+- Use the minecraft servers port in LAN worlds
+- Added Essential compatibility